[feature-81] Move layout under admin
