+++ src/admin/layout/header/bar/index.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+import { Icon } from 'admin/layout/header/icon'
+import { Wrapper } from 'admin/layout/header/bar/wrapper'
+import { DesktopButtons } from 'admin/layout/header/desktop-buttons'
+
+export const Bar = () => <Wrapper>
+  <Icon />
+  <DesktopButtons />
+</Wrapper>
