+++ src/admin/layout/header/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
 
-import { Bar } from 'layout/admin/header/bar'
+import { Bar } from 'admin/layout/header/bar'
 
 export const Header = () => <Bar />
