--- src/layout/admin/header/bar/index.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-import { Icon } from 'layout/admin/header/icon'
-import { Wrapper } from 'layout/admin/header/bar/wrapper'
-import { DesktopButtons } from 'layout/admin/header/desktop-buttons'
-
-export const Bar = () => <Wrapper>
-  <Icon />
-  <DesktopButtons />
-</Wrapper>
